requirements.txt
psycopg[binary] 
psycopg2-binary 
asyncpg 
python-dotenv
langchain
langgraph
fastapi